requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
google-generativeai